move gopassutils and kvm to pkg
