+++ GitlabInstance.go
@@ -7,9 +7,9 @@ import (
 	gitlab "gitlab.com/gitlab-org/api/client-go"
 
 	"github.com/asciich/asciichgolangpublic/datatypes/stringsutils"
-	"github.com/asciich/asciichgolangpublic/gopass"
 	"github.com/asciich/asciichgolangpublic/logging"
 	"github.com/asciich/asciichgolangpublic/parameteroptions"
+	"github.com/asciich/asciichgolangpublic/pkg/gopassutils"
 	"github.com/asciich/asciichgolangpublic/tracederrors"
 	"github.com/asciich/asciichgolangpublic/urlsutils"
 )
@@ -180,7 +180,7 @@ func (g *GitlabInstance) Authenticate(authOptions *GitlabAuthenticationOptions) 
 	}
 
 	for _, gopassPath := range authOptions.AccessTokensFromGopass {
-		credentialExists, err := gopass.CredentialExists(gopassPath)
+		credentialExists, err := gopassutils.CredentialExists(gopassPath)
 		if err != nil {
 			return err
 		}
@@ -197,7 +197,7 @@ func (g *GitlabInstance) Authenticate(authOptions *GitlabAuthenticationOptions) 
 
 		getSecretOptions := parameteroptions.NewGopassSecretOptions()
 		getSecretOptions.SetSecretPath(gopassPath)
-		accessToken, err := gopass.GetCredentialValueAsString(getSecretOptions)
+		accessToken, err := gopassutils.GetCredentialValueAsString(getSecretOptions)
 		if err != nil {
 			return err
 		}
