+++ pkg/kvm/KvmRemoveVmOptions.go
